// src/app/core/models/catalogos/servicios.model.ts
export interface Servicio {
    servicioId: number;
    nombre: string;
    precio: number;
    categoriaId: number;
    categoriaNombre: string;
    duracion: number;
    descripcion?: string;
    publicado: boolean;
    estado: boolean;
    urlsMultimedia: string[];
}

export interface ServicioRequest {
    nombre: string;
    precio: number;
    categoriaId: number;
    duracion: number;
    descripcion?: string;
    publicado: boolean;
    estado: boolean;
}

export interface ServicioFiltro {
    page?: number;
    size?: number;
    sort?: string;
    nombre?: string;
    categoriaId?: number;
    precioMin?: number;
    precioMax?: number;
    publicado?: boolean;
    estado?: boolean;
}